requirements.txt
ccxt
pandas
numpy
pandas-ta
python-telegram-bot>=20.0
aiohttp
python-dotenv
requests
schedule
